Added support for new "Position" parameter. Fixed a NullPointerException with "Show Controls"="no". Documented all supported parameters.
git-svn-id: svn://svn.code.sf.net/p/tigervnc/code/trunk@2512 3789f03b-4d11-0410-bbf8-ca57d06f2519

+++ java/src/com/tightvnc/rfbplayer/FbsInputStream.java
@@ -116,7 +116,7 @@ class FbsInputStream extends InputStream {
     return timeOffset;
   }
 
-  public synchronized void setTimeOffset(int pos)
+  public synchronized void setTimeOffset(long pos)
   {
     // FIXME: Seeking works only in paused mode.
     paused = true;
